api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
langchainhub
langchain-classic  # Fallback for AgentExecutor in LangChain 1.0+
supabase
openai
pydantic
langfuse
pypdf
python-docx
unstructured
pandas
openpyxl
pymupdf
mistralai
sentry-sdk[fastapi]

